Move single-page components to the respective route folder #120
