+++ src/routes/animation_switcher.svelte
