+++ src/routes/menu_items_sub.svelte
@@ -1,9 +1,9 @@
 <script>
 	import { is_min_width_768 } from '$lib/stores'
 	import AnimationSwitcher from './animation_switcher.svelte'
-	import GithubLink from './github_link.svelte'
-	import ThemeSwitcher from './theme_switcher.svelte'
-	import TwitterLink from './twitter_link.svelte'
+	import GithubLink from '../lib/components/github_link.svelte'
+	import ThemeSwitcher from '../lib/components/theme_switcher.svelte'
+	import TwitterLink from '../lib/components/twitter_link.svelte'
 	import VolumeSwitcher from './volume_switcher.svelte'
 </script>
